requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiohttp-retry==2.9.1
aiosignal==1.4.0
amqp==5.3.1
asgiref==3.9.1
attrs==25.3.0
awscli==1.40.0
billiard==4.2.1
black==24.10.0
boto3==1.38.1
botocore==1.38.1
cachetools==5.5.2
celery==5.4.0
certifi==2025.1.31
cffi==1.17.1
channels==4.3.1
channels_redis==4.3.0
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
crispy-bootstrap4==2024.10
cron-descriptor==1.4.5
cryptography==44.0.1
defusedxml==0.7.1
Django==5.1.6
django-allauth==65.4.1
django-appconf==1.1.0
django-celery==3.1.17
django-celery-beat==2.8.1
django-ckeditor==6.7.2
django-cors-headers==4.7.0
django-countries==7.6.1
django-country==0.1.0
django-crispy-forms==2.3
django-cryptography==1.1
django-environ==0.12.0
django-extensions==4.1
django-filter==25.1
django-jazzmin==3.0.1
django-js-asset==3.0.1
django-otp==1.6.1
django-ranged-response==0.2.0
django-ratelimit==4.1.0
django-recaptcha==4.1.0
django-redis==6.0.0
django-simple-captcha==0.6.2
django-storages==1.14.6
django-taggit==6.1.0
django-timezone-field==7.1
django-user-agents==0.4.0
django_celery_results==2.6.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
docutils==0.19
fpdf==1.7.2
frozenlist==1.7.0
geographiclib==2.0
geopy==2.4.1
google-api-core==2.24.1
google-api-python-client==2.161.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.68.0
gunicorn==23.0.0
httplib2==0.22.0
idna==3.10
jmespath==1.0.1
kombu==5.4.2
msgpack==1.1.1
multidict==6.6.3
mypy-extensions==1.0.0
mysqlclient==2.2.7
numpy==2.2.5
oauthlib==3.2.2
packaging==25.0
pathspec==0.12.1
paypalrestsdk==1.13.3
phonenumbers==9.0.3
pillow==11.1.0
platformdirs==4.3.6
prompt_toolkit==3.0.50
propcache==0.3.2
proto-plus==1.26.0
protobuf==5.29.3
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycountry==24.6.1
pycparser==2.22
PyJWT==2.9.0
pyOpenSSL==25.0.0
pyparsing==3.2.1
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python3-openid==3.2.0
pytz==2025.1
PyYAML==6.0.2
qrcode==8.2
redis==6.2.0
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
s3transfer==0.12.0
setuptools==80.9.0
shortuuid==1.0.13
six==1.17.0
social-auth-app-django==5.4.3
social-auth-core==4.5.6
sqlparse==0.5.3
stripe==11.5.0
twilio==9.7.0
typing_extensions==4.12.2
tzdata==2025.1
ua-parser==1.0.1
ua-parser-builtins==0.18.0.post1
uritemplate==4.1.1
urllib3==2.3.0
user-agents==2.2.0
validator==0.7.1
vine==5.1.0
wcwidth==0.2.13
wheel==0.45.1
whitenoise==6.9.0
yarl==1.20.1
